+++ rasterx/stroke.go
@@ -182,8 +182,8 @@ func strokeArc(p Adder, a, s1, s2 fixed.Point26_6, clockwise bool, trimStart,
 	// or cubic Bezier curves", 2003
 	// https://www.spaceroots.org/documents/elllipse/elliptical-arc.pdf
 	// The method was simplified for circles.
-	theta1 := math.Atan2(float64(s1.Y-a.Y), float64(s1.X-a.X))
-	theta2 := math.Atan2(float64(s2.Y-a.Y), float64(s2.X-a.X))
+	theta1 := math.Atan2(float32(s1.Y-a.Y), float32(s1.X-a.X))
+	theta2 := math.Atan2(float32(s2.Y-a.Y), float32(s2.X-a.X))
 	if !clockwise {
 		for theta1 < theta2 {
 			theta1 += math.Pi * 2
@@ -195,27 +195,27 @@ func strokeArc(p Adder, a, s1, s2 fixed.Point26_6, clockwise bool, trimStart,
 	}
 	deltaTheta := theta2 - theta1
 	if trimStart > 0 {
-		ds := (deltaTheta * float64(trimStart)) / float64(1<<tStrokeShift)
+		ds := (deltaTheta * float32(trimStart)) / float32(1<<tStrokeShift)
 		deltaTheta -= ds
 		theta1 += ds
 	}
 	if trimEnd > 0 {
-		ds := (deltaTheta * float64(trimEnd)) / float64(1<<tStrokeShift)
+		ds := (deltaTheta * float32(trimEnd)) / float32(1<<tStrokeShift)
 		deltaTheta -= ds
 	}
 
 	segs := int(math.Abs(deltaTheta)/(math.Pi/cubicsPerHalfCircle)) + 1
-	dTheta := deltaTheta / float64(segs)
+	dTheta := deltaTheta / float32(segs)
 	tde := math.Tan(dTheta / 2)
 	alpha := fixed.Int26_6(math.Sin(dTheta) * (math.Sqrt(4+3*tde*tde) - 1) * (64.0 / 3.0)) // Math is fun!
-	r := float64(Length(s1.Sub(a)))                                                        // Note r is *64
+	r := float32(Length(s1.Sub(a)))                                                        // Note r is *64
 	ldp := fixed.Point26_6{X: -fixed.Int26_6(r * math.Sin(theta1)), Y: fixed.Int26_6(r * math.Cos(theta1))}
 	ds1 = ldp
 	ps1 = fixed.Point26_6{X: a.X + ldp.Y, Y: a.Y - ldp.X}
 	firstPoint(ps1)
 	s1 = ps1
 	for i := 1; i <= segs; i++ {
-		eta := theta1 + dTheta*float64(i)
+		eta := theta1 + dTheta*float32(i)
 		ds2 = fixed.Point26_6{X: -fixed.Int26_6(r * math.Sin(eta)), Y: fixed.Int26_6(r * math.Cos(eta))}
 		ps2 = fixed.Point26_6{X: a.X + ds2.Y, Y: a.Y - ds2.X} // Using deriviative to calc new pt, because circle
 		p1 := s1.Add(ldp.Mul(alpha))
